GLTF: merge GLTFcollisionmodel and GLTFmodel together, introduce a collisions bool
