+++ src/Shapes/GLTFModel.tsx
@@ -1,29 +1,57 @@
-import { useLoader } from '@react-three/fiber';
+import React, { useMemo } from 'react';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader';
-import React from 'react';
+import { useLoader } from '@react-three/fiber';
+import { Box3 } from 'three';
+import { BoxProps, useBox } from '@react-three/cannon';
+import { Vector3 } from 'three';
+import { SkeletonUtils } from 'three-stdlib';
 
-type ModelProps = {
-  modelPath: string;
-  scaleFactor: number;
-  position: Array<number>;
-};
-/**
- * Model is a simple wrapper for GLTFLoader.
- * @param props
- * modelPath -> Path to gltf or glb folder or file. Paths can either be local or reference an S3 bucket.
- * scaleFactor -> Factor to scale the x,y,z of the gltf.scene object.
- * position -> Array of x,y,z coordinates to define the primitive's position in the scene. This is a small abstraction for
- *  the native position property present in scene objects, as primitive elements require their position to be directly modified.
- *
- * @returns model to render
- */
-function GLTFModel(props: ModelProps) {
-  const gltf = useLoader(GLTFLoader, props.modelPath);
-  gltf.scene.scale.multiplyScalar(props.scaleFactor);
-
-  return (
-    <primitive position={props.position} object={gltf.scene} dispose={null} />
-  );
+interface CustomGLTF extends BoxProps{
+  modelScaleFactor: number
+  bboxScaleFactor: number
+  modelPath: string
+  collision: boolean
 }
 
-export default GLTFModel;
+/**
+ * GLTFCollisionModel component renders a GLTF with a bounding box for collision detection.
+ * 
+ * This component accepts props that determine its size, position, type, 
+ * a callback to be called upon a collision, path to GLTF to load, and scaling factors for the model and bbox.
+ * 
+ * The component loads a GLTF with the useLoader hook, makes memoized calls to clone the loaded scene
+ * in order to be able to load multiple instances of the same GLTF, then scales the model and bounding box
+ * to the provided props.  The scaled bounding box hooks into the Cannon physics engine with the useBox hook and returned collisionRef. 
+ * @param props customGLTF
+ * @returns
+ */
+export default function GLTFModel(props: CustomGLTF) {
+  const { scene } = useLoader(
+    GLTFLoader,
+    props.modelPath,
+  );
+  // Must use Skeleton utils to support copies of skinned meshes https://github.com/mrdoob/three.js/issues/11573
+  let copiedScene = useMemo(() => SkeletonUtils.clone(scene), [scene]);
+  copiedScene.scale.multiplyScalar(props.modelScaleFactor);
+  let collisionRef;
+  if (props.collision){
+    let bbox = useMemo(() => new Box3().setFromObject(copiedScene), [copiedScene]);
+    const bboxSize = bbox.getSize(new Vector3());
+    const scaledBbox = bboxSize.multiplyScalar(props.bboxScaleFactor).toArray();
+    collisionRef = useBox(() => ({
+      args: scaledBbox, //Must accept array and not vector3
+      mass: props.mass,
+      position: props.position,
+      type: props.type,
+      onCollide: props.onCollide,
+    }));
+  }
+ 
+
+  return (
+    <> 
+    { props.collision ? <primitive ref={collisionRef} position={props.position} object={copiedScene} dispose={null} /> :   <primitive position={props.position} object={copiedScene} dispose={null} />}
+
+    </>
+  );
+}
